switch to Int8 where needed

+++ modules/elastoplasticity/ElementMatrix.h
@@ -90,8 +90,8 @@ RealMatrix<6, 6> FemModuleElastoplasticity::_computeElementMatrixTria3(Cell cell
   RealMatrix<3, 3> C_tang;
 
   Int8 iGP = 0;
-  for (Int32 ix = 0; ix < 3; ++ix) {
-    for (Int32 iy = 0; iy < 3; ++iy) {
+  for (Int8 ix = 0; ix < 3; ++ix) {
+    for (Int8 iy = 0; iy < 3; ++iy) {
       C_tang(ix, iy) = m_C_tang_gp(cell, iGP, ix, iy);
     }
   }
@@ -223,8 +223,8 @@ RealMatrix<12, 12> FemModuleElastoplasticity::_computeElementMatrixTetra4(Cell c
 
   RealMatrix<6, 6> C_tang_3d;
   Int8 iGP = 0;
-  for (Int32 ix = 0; ix < 6; ++ix) {
-    for (Int32 iy = 0; iy < 6; ++iy) {
+  for (Int8 ix = 0; ix < 6; ++ix) {
+    for (Int8 iy = 0; iy < 6; ++iy) {
       C_tang_3d(ix, iy) = m_C_tang_gp(cell, iGP, ix, iy);
     }
   }
